Split V2 implementation into separate ConvertToJsonCommandV2.cs file
